+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Inventory` tab to view items
 - `Signature` definition for `ClientVersion`, which allows version to be detected by signature bytes instead of hash
 - `ExecutableName` and `WindowClassName` properties for `ClientVersion` to support other clients